workflows/vlm_alerts/requirements.txt
notebook
opencv-python
requests
pillow
numpy
flask
gradio
streamlit
matplotlib 

